zshrc: replace incarg with improved version
My new version is more Do-What-I-Mean:
- If the cursor is on a number, increment that number.
- Otherwise, if there is a number anywhere to the right of the cursor,
  increment the leftmost such number.
- Otherwise, if there is a number anywhere to the left of the cursor,
  increment the rightmost such number.

The first two cases operate just like Vim.  The third is a feature so that if
your cursor is at the end of the line, you can increment the last number in
the line.

I have bound this to ^X^A / ^X^X to be more vim-like.