xfce4-terminal: fix opaque cursor
There is a bug [1] since 2010 that xfce4-terminal draws the cursor as an
opaque color rather than using the "invert" style.  This means that it is
impossible to see what's under the cursor.  The workaround, as noted in [1],
is to manually set CursorColor to the empty string in the rc file.

[1] https://bugs.launchpad.net/ubuntu/+source/xfce4-terminal/+bug/673489