vimrc: remove gitgutter sign colors, update line colors
The latest version of gitgutter automatically uses the right sign column
background color so we don't need to define it ourselves.  The line colors are
still too bright (those of diff) so we now use some very subtle line colors
that match our background color.

Here is the thread describing the bug referenced in the comment:
https://www.mail-archive.com/vim_use@googlegroups.com/msg45293.html