lircrc: use named sink in volume control commands
If there is more than one sink (output device), then sink #0 might not
always be the same on every reboot.  In my case, I added HDMI output,
and sometimes this is #0 instead of my PCI SoundBlaster card.  By using
the named sink, the volume buttons will always control the PCI output
and not the HDMI output.