vimrc: set 'syn sync fromstart' for Markdown
This fixes the problem that Markdown syntax gets messed up sometimes
when the size of the block is large, e.g. for a larged fenced code
block.

This also removes the old 'syn sync maxlines' command since that
actually did nothing; sync is reset when the syntax file is reloaded.