crawlrc: replace run with safe walk
This moves a single tile.  There is unfortunately no way to have both
run and safe walk bound at the same time, unless I use some obscure
keys.  The reason is that only bindkey can map keys to functions, but
bindkey does not accept ALT, SHIFT-CTRL, or any of the numpad keys.
Therefore, I either have to replace the vi keys with safe walk or find
some other keys in which to stick safe walk.

To get shift-numpad to work, I have to use a macro because on my version
of tiles, these keys are bound directly as opposed to being aliases as
in some other versions of tiles (e.g. web tiles, I think).