zshrc: show hostname in prompt if under SSH
Previously I used $MY_MACHINE to determine whether or not to set the
hostname in the prompt, since I often SSH'd to a specific machine. But
that ends up being a pain to configure. Almost always I want it to show
the hostname if and only if SSHing, so that is a much saner default.